server/requirements.txt
PyYAML>=6.0
sqlparse>=0.5.0

git+ssh://devops.aishu.cn:22/AISHUDevOps/One-Architecture/_git/proton-rds-sdk-py@release/2.0.0
